renamed property_file -> property_type
